refactor: rename SocialMediaLinks-styling file
